requirements.txt
tensorflow>=2.5.0
scikit-learn
streamlit
pandas
numpy
matplotlib
seaborn
opencv-python-headless
flask
Pillow

